changed save file index type from uint to int for functional indexing in BackupManager, Update version to 5.8.1

+++ src/managers/backupmgr/cleanup.go
@@ -196,7 +196,7 @@ func (m *BackupManager) getBackupSaveFiles() ([]BackupSaveFile, error) {
 	})
 	// Add the index to each save
 	for i := range saves {
-		saves[i].Index = uint(i)
+		saves[i].Index = i
 	}
 
 	return saves, nil
